FIX: warnings, signed v.s unsigned comparison

+++ Testing/Code/BasicFilters/itkNarrowBandImageFilterBaseTest.cxx
@@ -68,8 +68,8 @@ namespace itk
       typename ImageType::IndexType tl= this->GetInput()->GetRequestedRegion().GetIndex();
       typename Superclass::IndexType in;
 
-      for (in[0]=32+tl[0];in[0]<tl[0]+sz[0];in[0]++)
-        for (in[1]=tl[1]+32;in[1]<tl[1]+sz[1];in[1]++)       
+      for (in[0]=32+tl[0];in[0]<tl[0]+(long int)(sz[0]);in[0]++)
+        for (in[1]=tl[1]+32;in[1]<tl[1]+(long int)(sz[1]);in[1]++)       
           this->InsertNarrowBandNode (in); 
      m_RegionList=m_NarrowBand->SplitBand(this->GetNumberOfThreads());
     }
